Add files via upload
Agrego Jugador basico con primer test

+++ src/modelo/herramientas/Herramienta.java
@@ -21,7 +21,7 @@ public abstract class Herramienta{
 
     }
 
-    public int getDurabilidad(){
+    public float getDurabilidad(){
 
         return this.durabilidad;
